Enable running HLT secondary vertexing on tracks including validation

+++ SimTracker/VertexAssociation/python/secondaryVertexAssociatorByPositionAndTracks_cfi.py
@@ -3,9 +3,9 @@ import FWCore.ParameterSet.Config as cms
 # for SV of type reco::Vertex
 secondaryVertexAssociatorByPositionAndTracks = cms.EDProducer("VertexAssociatorByPositionAndTracksProducer",
     absZ = cms.double(1.0),
-    sigmaX = cms.double(5),
-    sigmaY = cms.double(5),
-    sigmaZ = cms.double(5),
+    sigmaX = cms.double(10),
+    sigmaY = cms.double(10),
+    sigmaZ = cms.double(10),
     maxRecoZ = cms.double(1000),
     absT = cms.double(-1),
     sigmaT = cms.double(-1),
